feat(docker): introduce InspectCommand trait for container inspection
Add the InspectCommand trait to encapsulate the logic for inspecting
Docker containers, improving code modularity and reusability. This
change removes the inspect method from DockerClient, replacing it with
the trait, which simplifies the DockerClient class and centralizes
inspection logic. Additionally, a new unit test, InspectCommandTest,
is introduced to ensure the functionality of the InspectCommand trait.
Unused imports are removed from various Docker output classes to
clean up the codebase.
