+++ src/Docker/DockerClient.php
@@ -4,18 +4,16 @@ namespace Testcontainers\Docker;
 
 use Symfony\Component\Process\Process;
 use Testcontainers\Docker\Command\BaseCommand;
+use Testcontainers\Docker\Command\InspectCommand;
 use Testcontainers\Docker\Command\RunCommand;
 use Testcontainers\Docker\Exception\DockerException;
-use Testcontainers\Docker\Exception\InvalidValueException;
 use Testcontainers\Docker\Exception\NoSuchContainerException;
-use Testcontainers\Docker\Exception\NoSuchObjectException;
-
 use Testcontainers\Docker\Output\DockerFollowLogsOutput;
-use Testcontainers\Docker\Output\DockerInspectOutput;
 use Testcontainers\Docker\Output\DockerLogsOutput;
 use Testcontainers\Docker\Output\DockerNetworkCreateOutput;
 use Testcontainers\Docker\Output\DockerProcessStatusOutput;
 use Testcontainers\Docker\Output\DockerStopOutput;
+
 use function Testcontainers\array_flatten;
 
 /**
@@ -28,46 +26,7 @@ class DockerClient
 {
     use BaseCommand;
     use RunCommand;
-
-    /**
-     * Inspect a Docker container.
-     *
-     * This method wraps the `docker inspect` command to retrieve detailed information about the specified container.
-     *
-     * @param string $containerId The ID of the container to inspect.
-     * @return DockerInspectOutput The output of the Docker inspect command, including detailed information about the container.
-     *
-     * @throws DockerException If the Docker command fails for any other reason.
-     * @throws NoSuchObjectException If the specified container does not exist.
-     * @throws InvalidValueException If the output of the `docker inspect` command is not valid JSON.
-     */
-    public function inspect($containerId)
-    {
-        $commandline = array_filter(array_flatten([
-            $this->command,
-            $this->arrayToArgs($this->options),
-            'inspect',
-            $containerId,
-        ]));
-        $process = new Process(
-            $commandline,
-            $this->cwd,
-            $this->env,
-            $this->input,
-            $this->timeout,
-            $this->proc_options
-        );
-        $process->run();
-
-        if (!$process->isSuccessful()) {
-            if (NoSuchObjectException::match($process->getErrorOutput())) {
-                throw new NoSuchObjectException($process);
-            }
-            throw new DockerException($process);
-        }
-
-        return new DockerInspectOutput($process);
-    }
+    use InspectCommand;
 
     /**
      * Stop one or more running Docker containers.
