+++ src/Docker/Output/DockerLogsOutput.php
@@ -2,8 +2,6 @@
 
 namespace Testcontainers\Docker\Output;
 
-use Testcontainers\Docker\Output\DockerOutput;
-
 /**
  * Represents the output of a Docker `logs` command executed via Symfony Process.
  *
