+++ src/Docker/Output/DockerNetworkCreateOutput.php
@@ -2,8 +2,6 @@
 
 namespace Testcontainers\Docker\Output;
 
-use Testcontainers\Docker\Output\DockerOutput;
-
 class DockerNetworkCreateOutput extends DockerOutput
 {
 }
