+++ src/Docker/Output/DockerRunOutput.php
@@ -2,8 +2,6 @@
 
 namespace Testcontainers\Docker\Output;
 
-use Testcontainers\Docker\Output\DockerOutput;
-
 /**
  * Represents the output of a Docker `run` command executed via Symfony Process.
  *
