+++ src/Docker/Output/DockerRunWithDetachOutput.php
@@ -3,7 +3,6 @@
 namespace Testcontainers\Docker\Output;
 
 use Symfony\Component\Process\Process;
-use Testcontainers\Docker\Output\DockerRunOutput;
 use Testcontainers\Docker\Types\ContainerId;
 
 /**
