+++ src/Docker/Output/DockerStopOutput.php
@@ -3,7 +3,6 @@
 namespace Testcontainers\Docker\Output;
 
 use Symfony\Component\Process\Process;
-use Testcontainers\Docker\Output\DockerOutput;
 
 /**
  * Handles the output of a Docker `stop` command executed via Symfony Process.
